Update trace validation to handle multiple ? for all platforms (#423)
*Issue description:*
Follow up from this PR:
https://github.com/aws-observability/aws-application-signals-test-framework/pull/419

The issue is related to cw agent so all platforms are affected. Adding
the REGEX fix to all dotnet validation templates except for adot sigv4
since it does not use cw agent

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.