[Genesis] Release Testing - Validator Changes (#425)
*Description of changes:*

This PR is part 1 of the release testing changes to support the incoming
release of the latest Genesis/Agentic AI Observability changes done in
ADOT Python which contains all of the changes to `validator`

We need to validate that we are correctly emitting all 3 telemetry
signals to CW backend.

1. Expands `Context` class to also take in a given traceId - as a part
of the Agent Observability, we've split off user input/output prompts
from Gen AI span events as a separate log event. The idea is that users
will be able to rebuild their entire Gen AI span event by querying for
logs with the same trace id in `aws/spans` and a custom log group. As a
part of this testing we need to verify that trace Id is being propagated
from their `Agentic AI application + ADOT --> CW` correctly and can be
linked together.

2. Updated `AWS OTLP log filter` patterns to capture and validate GenAI
input/output prompt log events.

3. Updated `TraceValidator` to validate pre-generated trace IDs for
GenAI release testing

4. Updated `MetricValidator` to handle GenAI metrics special case: ADOT
Python captures all OTel metrics from the Agentic AI application
Instrumentor and converts them to EMF format. Validation for this path
just checks for any metric emission to a custom namespace since the
exact metric structure is unknown

*Ensure you've run the following tests on your changes and include the
link below:*
Example of a successful run of this test:
https://github.com/liustve/aws-application-signals-test-framework/actions/runs/16172487807/job/45649328445

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.