[Genesis] Release Testing - Workflow Changes (#426)
This PR is part 2 of the release testing changes to support the incoming
release of the latest Genesis/Agentic AI Observability changes done in
ADOT Python which contains all of the changes for adding the release
workflow including sample app, terraform files, and GitHub workflow.

*Description of changes:*

- Added `python-ec2-genai-test.yml` workflow for automated GenAI service
testing which contains a step for generating a random X-Ray trace ID and
W3c trace ID to validate trace context propagation and Gen AI log event
+ span linking

- Added a simple Python GenAI service `server.py` created with LangChain
and instrumented with LangChain instrumentor and ADOT Python.

- Added generate_traffic.sh script for generating traffic to hit the
`/ai-chat` endpoint to generate traces `server.py`

- `Logs`: Validates Gen AI log event appears in in CloudWatch

- `Traces`: Validates trace propagation and Gen AI log event + span
linking

- `Metrics`: Validates that OTel metrics emitted by the sample
application are converted to EMF metrics and are able to be published to
CloudWatch

Successful run:
https://github.com/aws-observability/aws-application-signals-test-framework/actions/runs/16183070365/job/45683540103

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.