Append SystemCGroup from containderd config file (#435)
*Issue description:*
There was a change in the most recent EC2 AMI such that the config file
for containderd has the attribute `SystemdCgroup` missing. (Before the
attribute was present with the value set to `SystemdCgroup = false`)
This attribute needs to be set to true in order for the k8s cluster to
work.

*Description of changes:*
The sed command will initially replace the false value with true for
SystemCgroup. If the systemCgroup doesn't exist though, it will try to
append the attribute to the
`[plugins.'io.containerd.cri.v1.runtime'.containerd.runtimes.runc.options]
` plugin

Ran the script on personal account and verified that k8s is running
again
By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.