[Java] Add API for testing downstream error code (#436)
## Changes
Adding a new API `/status/{code}?ip=...` that calls the downstream
service similar to `/remote-service` and specifies the code it wants the
remote service to respond with. The remote service now accepted
`/status/{code}` and responds with status code `{code}`. The main
service will respond with OK and the trace ID in any case.

## Testing
Built and ran the applications locally, confirming functionality of the
new APIs.

## Risks
New API does not change any functionality of old ones. Could cause build
issues, but we use other imports from the same dependencies and the same
logic in `/remote-service` so there is little to no risk.

## Rollback procedure

Revert commit and rerun sample app deployment script.


#### By submitting this pull request, I confirm that my contribution is
made under the terms of the Apache 2.0 license.