updating 3p actions from version ID to commit SHA (#456)
This pr updates all 3p actions from using version ID tags to being
pinned with their commit SHA.

Git revert is safe unless SHA pinning is enforced for this repository in
the settings. In that case all pinned commit SHA must be corresponding
to the version tag that was specified before this pr. Alternatively
functionality and/or arguments of commands must be updated to satisfy
changes between the versions.

References:
https://github.com/actions/checkout
https://github.com/actions/setup-python
https://github.com/actions/setup-java
https://github.com/actions/setup-node
https://github.com/actions/cache
https://github.com/actions/upload-artifact
https://github.com/aws-actions/configure-aws-credentials  
https://github.com/actions/download-artifact
https://github.com/aws-actions/aws-secretsmanager-get-secrets

https://github.com/gradle/actions/blob/f8140229023a7015c7ce4df6f7c390a3cace8f83/docs/deprecation-upgrade-guide.md#using-the-action-to-execute-gradle-via-the-arguments-parameter-is-deprecated
https://github.com/hashicorp/setup-terraform
https://github.com/ruby/setup-ruby
https://github.com/aws-actions/amazon-ecr-login
https://github.com/actions/setup-dotnet

NOTE: TESTS RUNNING ON A SINGLE EKS CLUSTER CANNOT BE RUN IN PARALLEL.
See the
[needs](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idneeds)
keyword to run tests in succession.
- Run Java EKS on `e2e-playground` in us-east-1 and eu-central-2
- Run Python EKS on `e2e-playground` in us-east-1 and eu-central-2
- Run metric limiter on EKS cluster `e2e-playground` in us-east-1 and
eu-central-2
- Run EC2 tests in all regions
- Run K8s on a separate K8s cluster (check IAD test account for master
node endpoints; these will change as we create and destroy clusters for
OS patching)

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: Thomas Pierce <thp@amazon.com>