[Fix] Add setup-java action to Java Lambda Layer performance tool workflow
Currently, the workflow does not set up Java explicitly, so it always uses the default Java version preinstalled on Linux to build the Java Lambda Layer. This causes build failures because OTel/ADOT requires Java 21, while the default version on Linux is Java 17.

Although the workflow includes a java_version parameter, it is not currently being used.

This fix explicitly sets up Java to the version specified by the java_version parameter.