[ADOT Java Lambda Perf Tool]: Update the default ADOT-Java branch from release/v2.11.x to main (#477)
In the ADOT Java release MCM, the performance tool runs on the main
branch. This change ensures consistency between the default branch and
MCM.

Also, the default Java SDK version has been changed from 17 to 21 to
maintain alignment with both OTel and ADOT.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.