Added .NET Framework app as well as Windows IaC files for EC2, EKS, and ECS (#517)
*Description of changes:*
There are multiple changes in this PR. 
1. Added a new .NET Framework app that can be deployed using docker or
as a zip file where it will be built and deployed on a native EC2
windows instance.
2. Added a Dockerfile.windows in the already existing .NET ASP.NET Core
application to support deployment on windows EKS/ECS
3. Added new cdk-native-windows and terraform-native-windows that deploy
native (non-docker) windows EC2 instances for .NET applications
4. Updated EKS code to support windows containerized deployments
5. Added new cdk-windows and terraform-windows that deploys
containerized windows application to ECS
6. Added a bunch of new enablement tasks:
- Added new EC2 enablement tasks for .NET applications deployed on
native EC2
- Added new ECS enablement tasks for .NET applications deployed on ECS
Windows
- Didn't add anything new to EKS since the same .NET enablement task
applies to both windows and linux EKS.

*Notes:*
- .NET Framework enablement isn't available on EKS/ECS. It requires IIS
to start the application which we don't support yet. The cloudwatch
addon on EKS doesn't and there are issue starting IIS on ECS when I was
doing some testing.
- .NET Framework enablement has been testing in Native EC2. The
application gets instrumented. However, the application breaks and it's
an issue with ADOT that we need to figure out. But for the purposes of
MCP Enablement, it's working.

*Testing:*
I ran the enablement tasks for EKS, ECS and EC2 windows and verified
using manual deployments to verify that ADOT is enabled and that
Application Signals was receiving metrics.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: Michael He <53622546+yiyuan-he@users.noreply.github.com>